requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
requests==2.28.0
google-auth==2.16.0
google-auth-oauthlib==0.5.2
google-auth-httplib2==0.1.0
jinja2==3.1.2
python-dotenv==0.21.1
